+++ restx-core/src/main/java/restx/endpoint/mappers/EndpointParameterMapper.java
@@ -1,7 +1,9 @@
-package restx.endpoint;
+package restx.endpoint.mappers;
 
 import restx.RestxRequest;
 import restx.RestxRequestMatch;
+import restx.endpoint.EndpointParamDef;
+import restx.endpoint.EndpointParameterKind;
 
 public interface EndpointParameterMapper {
     <T> T mapRequest(EndpointParamDef endpointParamDef, RestxRequest request, RestxRequestMatch match, EndpointParameterKind parameterKind);
