+++ restx-core/src/main/java/restx/endpoint/mappers/EndpointParameterMapperFactory.java
@@ -1,6 +1,7 @@
-package restx.endpoint;
+package restx.endpoint.mappers;
 
 import com.google.common.base.Optional;
+import restx.endpoint.EndpointParamDef;
 
 public interface EndpointParameterMapperFactory {
     Optional<? extends EndpointParameterMapper> getEndpointParameterMapperFor(EndpointParamDef endpointParamDef);
